ref(ngrx-ducks): move core/ to __deprecated__/
