+++ ngrx/ducks/src/lib/__deprecated__/create-reducer-from/create-reducer-from.spec.ts
