+++ ngrx/ducks/src/lib/__deprecated__/ducks/create-duck.spec.ts
@@ -6,7 +6,9 @@ describe('SelfDispatchingAction', () => {
     it('should dispatch the action with the given payload', () => {
       const type = '[Login] User entered login data';
       const caseReducer = (_state: string, _payload: string) => 'new state';
-      const dispatchMock = jest.fn(payload => {});
+      const dispatchMock = jest.fn(payload => {
+        /** */
+      });
       const store = { dispatch: dispatchMock };
 
       const wa = createWiredAction(type, caseReducer);
