+++ ngrx/ducks/src/lib/__deprecated__/ducks/create-duck.ts
