+++ ngrx/ducks/src/lib/__deprecated__/ducks/create-ducks.spec.ts
