+++ ngrx/ducks/src/lib/__deprecated__/ducks/create-effect-dispatcher.ts
