+++ ngrx/ducks/src/lib/__deprecated__/types/__internal__/action-with-payload.ts
