+++ ngrx/ducks/src/lib/__deprecated__/types/__internal__/action.ts
