+++ ngrx/ducks/src/lib/__deprecated__/types/action-creator/action-creator-without-payload.ts
