+++ ngrx/ducks/src/lib/__deprecated__/types/case-reducer/case-reducer-without-payload.ts
