+++ ngrx/ducks/src/lib/__deprecated__/types/ducks/duck-with-payload.ts
