+++ ngrx/ducks/src/lib/__deprecated__/types/ducks/duck-without-payload.ts
