+++ ngrx/ducks/src/lib/__deprecated__/types/ducks/ducks.ts
