+++ ngrx/ducks/src/lib/__deprecated__/types/method/method-taking-two-parameters.ts
