+++ ngrx/ducks/src/lib/__deprecated__/types/wired-actions/extract-type-from-wired-action.ts
