+++ ngrx/ducks/src/lib/__deprecated__/types/wired-actions/wired-action.ts
