+++ ngrx/ducks/src/lib/__deprecated__/types/wired-actions/wired-actions.ts
