+++ ngrx/ducks/src/lib/__deprecated__/wired-actions/create-wired-action.ts
