+++ ngrx/ducks/src/lib/__deprecated__/wired-actions/wire-up-actions.ts
