+++ ngrx/ducks/src/lib/ducks/create-duck.ts
@@ -1,5 +1,5 @@
 import { methodsFrom } from '../class';
-import { WiredActions } from '../core/types';
+import { WiredActions } from '../__deprecated__/types';
 import { extractWiredAction } from './extract-wired-action';
 
 export function createDuck<T extends new () => InstanceType<T>>(
