+++ ngrx/ducks/src/lib/typings/duck-service.ts
@@ -1,4 +1,4 @@
-import { PickSelector } from '../core/types';
+import { PickSelector } from '../__deprecated__/types';
 import { DuckActionDispatcher } from './duck-action-dispatcher';
 
 export type Duck<T> = { [K in keyof T]: DuckActionDispatcher<T[K]> } &
